refactor: extract the shared phrase-set form state into a hook
CreatePhraseSetDialog and EditPhraseSetDialog declared the same
nineteen pieces of state and carried byte-identical copies of
loadGroups, loadPhrases, handlePhraseToggle, the phrase filter
predicate and the availableCategories derivation. All of that now
lives in usePhraseSetForm, along with resetForm and a
buildAccessPayload helper for the access/expiry half of the submit
body.

buildAccessPayload reproduces the original shape exactly, including
sending undefined rather than [] when no groups or usernames are
selected - the backend distinguishes "not supplied" from "explicitly
empty", so collapsing those would have changed behaviour.

The step JSX is deliberately left alone. It looks duplicated but is
not: the create dialog hides the options that don't apply to
crosswords and pages its phrase list as you scroll, neither of which
the edit dialog does. Merging them would mean threading flags through
a shared component to switch those behaviours back off, which is the
same conditional sprawl that argued against a single mode="create"
component in the first place.

Both dialogs also stop taking a token prop, since the API hooks they
use no longer need one.

Verified in a browser end to end: created a set with three selected
phrases and reopened it in the edit dialog and saved, with no 4xx or
5xx. Checked the persisted row afterwards - name, access type,
auto-delete date, config and all three phrase links are correct.
